+++ src/app/page.tsx
@@ -71,7 +71,7 @@ export default function Home() {
 
   const handleSubmit = (value: number) => {
     const battlePower = getBattlePower(value);
-    router.push(`/loding?score=${battlePower}`);
+    router.push(`/loading?score=${battlePower}`);
   };
 
   // 音源のオン設定
